Move the whole company dashboard section to a component
